Spring Intergration Добавил сервис Orders для его запуска требуется запустить проект spring-integration -> SpringIntegrationApplication В eshop добавили OrderServiceImpl И на странице Bucket добавили кнопку Commit bucket которая отвечает за функционал integrations, по ее нажатию если включен spring-integration в папку Orders сохраняется заказ

+++ spring-eshop/src/main/java/com/akhtyamov/springeshop/controllers/BucketController.java
@@ -6,12 +6,13 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import java.security.Principal;
 
 @Controller
-@RequestMapping("/bucket")
+//@RequestMapping("/bucket")
 public class BucketController {
     private final BucketService bucketService;
 
@@ -19,7 +20,7 @@ public class BucketController {
         this.bucketService = bucketService;
     }
 
-    @GetMapping
+    @GetMapping("/bucket")
     public String aboutBucket(Model model, Principal principal){
         if (principal==null){
             model.addAttribute("bucket", new BucketDTO());
@@ -32,7 +33,17 @@ public class BucketController {
 
     @GetMapping("{id}/remove")
     public String remove(@PathVariable Long id, Principal principal){
-        bucketService.removeProduct(id,principal);
+        if (principal!=null) {
+            bucketService.removeProduct(id, principal);
+        }
+        return "redirect:/bucket";
+    }
+
+    @PostMapping("/bucket")
+    public String commitBucket(Principal principal){
+        if (principal!=null){
+            bucketService.commitBucketToOrder(principal.getName());
+        }
         return "redirect:/bucket";
     }
